server/face_service/requirements.txt
flask>=3.0
flask-cors>=4.0
deepface>=0.0.90
opencv-python-headless>=4.8
numpy>=1.24
pillow>=10.0
